requirements.txt
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.1
drf-nested-routers==0.90.0
